refactor(token): replace Token with ValidToken and Token interface
